Add string-of-int lint, basically go vet 1.15
This finds 21 failures, though only 3 of which are outside of tests.
In practice, these look like:
```go
# service/history/shard/controller.go:381:55: [string-of-int] dubious convertion of an integer into a string, use strconv.Itoa
info, err := c.GetHistoryServiceResolver().Lookup(string(shardID))
```
And they are *nearly always* a bug.

The correlation is so strong, `go test` on 1.15 and newer will by default refuse to compile a package with this code.
They have a pretty good bit on this in the release notes: https://golang.org/doc/go1.15#vet
To copy a relevant snippet from that:
> `string(9786)` does not evaluate to the string `"9786"`; it evaluates to the string `"\xe2\x98\xba"`, or `"☺"`.
So yeah.  That's basically never what is intended.

If you are on 1.15 or newer, you can reproduce these with `go vet -stringintconv ./...`.
Or just run `go test` on an affected package, it'll error rather than run tests.

We should *absolutely* prevent *all* of these.  We can't rely on `go` though, since this repo targets 1.12, so I'm glad this lint exists.

---

There is one spot I *did not* change (behaviorally), because I'm not sure what doing so will do.
The `shardIDstr` in service/frontend/adminHandler.go:251 is used as a hash-ring lookup key.
Changing this will change how shards hash, which could cause issues when deployed.

So, for now at least, I've converted it to the equivalent that the Go github issue recommends: `string(rune(int))`.
This produces the same values as before.

---

Strangely, this misses a fairly trivial case that `go vet` catches:
```go
--- a/service/frontend/adminHandler.go
+++ b/service/frontend/adminHandler.go
@@ -248,7 +248,7 @@ func (adh *adminHandlerImpl) DescribeWorkflowExecution(
        }

        shardID := common.WorkflowIDToHistoryShard(request.Execution.WorkflowID, adh.numberOfHistoryShards)
-       shardIDstr := string(shardID)             // misses this one
+	shardIDstr := string(rune(shardID)) // ...
        shardIDForOutput := strconv.Itoa(shardID) // correct, not changed in this diff

        historyHost, err := adh.GetMembershipMonitor().Lookup(common.HistoryServiceName, shardIDstr)
```
This might be due to the variable being handled correctly once?  I dunno.
I haven't been able to reproduce this with a smaller test, unfortunately.
Thankfully, this repo is open source, so it can serve as the reproduction.