Attempt to make some racy tests more reliable (#3961)
Unfortunately these tests do not have any guaranteed way to sequence these things,
so I wrote down the best-effort strategies and used them to hopefully make the code a bit more clear on intent.

At the core, what this code needs is a way to simultaneously block on one action and unblock another, and then do the reverse later.
You can do that by e.g. pushing a message on a non-buffered channel...
... but that would require significant complicating changes to this code just for testing.

So it's inherently racy.  Throw stuff at the wall until it sticks, but try to be clear about the goals.