Detect when IDL submodule is not initialized, warn user (#4172)
This is largely a continuation of #4155 , and is related to / helps resolve #4162 .

When the IDL submodule doesn't have files in it, builds can fail in a variety of obtuse ways.
Currently it fails when running protoc, as there's a symlink from proto/ into idls/, e.g. causing error from #4155 :
```
uber/cadence/matching/v1/service.proto:213:12: "api.v1.TaskListPartitionMetadata" is not defined.
proto/public: warning: directory does not exist.
uber/cadence/api/v1/history.proto: File not found.
...
```

Since this needs to be fixed *before* running a `make ...`, as its files are used as target prerequisites,
it can't be auto-initialized by the makefile.  Instead, it now fails the build with a descriptive error,
and the user needs to resolve it before trying again.

---

While building and testing this I also noticed that the fake protoc/thrift binaries were sometimes surprising me,
so I added a warning about them to `make clean`.  They *might* be reasonable to automatically clean up as well,
but I'm leaving them in place for now, so e.g. a build process doesn't accidentally fake -> clean -> do a full build.